requirement.txt

ansi2html==1.8.0
appdirs==1.4.4
atomicwrites==1.4.0
autoviz==0.1.730
backports.weakref==1.0.post1
brotlipy==0.7.0
catboost==1.2
category-encoders==2.6.1
certifi==2023.5.7
clyent==1.2.1
constantly==15.1.0
cssselect==1.1.0
Cython

dash==2.11.1
dash-table==5.0.0
datashape==0.5.2
Deprecated==1.2.14
deprecation==2.1.0
emoji==2.6.0
et-xmlfile==1.1.0
fonttools==4.25.0
graphviz==0.20.1
holoviews==1.14.9
hvplot==0.7.3
importlib-metadata==6.8.0
inflection==0.5.1
joblib==1.3.1
jsonpointer==2.1
kaleido==0.2.1
lightgbm==4.0.0

mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
orjson==3.9.2
pandas-dq==1.28
patsy==0.5.3
pep8==1.7.1
Pillow==9.4.0
plotly-resampler==0.9.1
ply==3.11
pyamg==5.0.1
pycaret

PyDispatcher==2.0.5
pyls-spyder==0.4.0
pyod==1.1.0
PyQt5==5.15.7
pytest==7.1.2
python-lsp-jsonrpc==1.0.0
pywin32==305
queuelib==1.5.0
retrying==1.3.4
schemdraw==0.15
scikit-base==0.5.0
scikit-plot==0.3.7
scipy==1.10.0

sklearn-pandas==2.2.0
sktime==0.21.0
slicer==0.0.7

tbats
TBB==2021.10.0
textblob==0.17.1
torch==2.0.1
trace-updater==0.0.9.1
tsdownsample==0.1.2
webencodings==0.5.1
wincertstore==0.2
wordcloud==1.9.2
xgboost==1.7.6
xlrd==2.0.1
xxhash==3.2.0
yellowbrick
zict==2.1.0
zstandard==0.19.0
streamlit